dj13_diploma/requirements.txt
asgiref==3.3.1
attrs==20.3.0
coverage==5.3
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
iniconfig==1.1.1
model-bakery==1.2.1
packaging==20.4
Pillow==8.0.1
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
